requirements.txt
amadeus==11.0.0
certifi==2024.2.2
charset-normalizer==3.4.1
click==8.1.8
crewai==0.86.0
crewai-tools==0.17.0
dnspython==2.7.0
email-validator==2.2.0
flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pandas==2.2.3
pydantic==2.10.4
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
openai==1.58.1